The change made to the code is the addition of a new type alias GetFieldName which is defined as a generic type that takes a subtype of TableKey and extracts the corresponding FieldName.
The new GetFieldName type can be used to retrieve the field names of a specific table by passing the table name as a type argument. For example, GetFieldName<'users'> will return the type keyof typeof db['users']['fields'].

This change adds more type safety to the code by allowing the compiler to check that the requested field names actually belong to the specified table.