The changes made to the Table interface and the Metadata type involve the use of a generic type parameter T extends TableKey.
In the original code, the Table interface used FieldName[TableKey] to specify the field names for each table. This means that the field names for a table were determined by the TableKey type, which is a union of all table names.

In the updated code, the Table interface is defined using a generic type parameter T extends TableKey. This means that the fields property of the Table interface now uses FieldName[T] to specify the field names for a specific table, based on the value of T.

The Metadata type is also updated to use the Table interface with the generic type parameter Table<TableKey>. This ensures that the fields property of each table in the metadata is defined using the correct set of field names for that table.

Overall, these changes make the Table interface and Metadata type more specific and type-safe, by using a generic type parameter to ensure that the correct set of field names is used for each table.